+++ src/card/cards/combat_constraints.rs
@@ -25,3 +25,4 @@ pub const RESET: CardDefinitionId = CardDefinitionId(1824);
 pub const LIVING_ARTIFACT: CardDefinitionId = CardDefinitionId(1825);
 pub const INSTILL_ENERGY: CardDefinitionId = CardDefinitionId(1826);
 pub const HOWLING_MINE: CardDefinitionId = CardDefinitionId(1827);
+pub const JADE_STATUE: CardDefinitionId = CardDefinitionId(1828);