+++ src/game/tests.rs
@@ -574,6 +574,7 @@ mod howling_mine;
 mod identity_and_mana;
 mod instill_energy;
 mod isd_dgm_deck_cards;
+mod jade_statue;
 mod laces;
 mod land_and_ability_layers;
 mod land_characteristics;